requirements.txt
streamlit==1.29.0
twilio==8.10.0
openai>=2.8.0
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
httpx>=0.28.0
